git-svn-id: http://svn.code.sf.net/p/shopizer/code/shopizer-portlet@2272 e80ee2c7-76f3-4005-aeaf-98396743f455

+++ sm-core/src/main/java/com/salesmanager/core/business/customer/service/attribute/CustomerOptionServiceImpl.java
@@ -61,9 +61,9 @@ public class CustomerOptionServiceImpl extends
 	public void delete(CustomerOption customerOption) throws ServiceException {
 		
 		//remove all attributes having this option
-		List<CustomerAttribute> attributes = customerAttributeService.getByCustomerOption(customerOption.getMerchantStore(), customerOption);
+		CustomerAttribute attribute = customerAttributeService.getByCustomerOptionId(customerOption.getMerchantStore(), customerOption.getId());
 		
-		for(CustomerAttribute attribute : attributes) {
+		if(attribute!=null) {
 			customerAttributeService.delete(attribute);
 		}
